Move DirIterator to io_util
